requirements.txt
duolingo-api
urwid